More refactoring and clean up of API server
